Create Lamp Task
To model a lamp as a Java object: make a Lampclass. 
This will contain at least one 
instance variablewhich is of typeBoolean
and ithold thestateof the lamp: i.e. Whetherit is ON or OFF.
 Addmethodsto do the following things: switch the light on and off, and check its current state, i.e., whether it is on or off.Next, write a launcher class with amain ()method to carry out the following tasks:
Create a lamp object;
Turn it ON and OFF;Print the lamp's on/off status to the console (i.e., usingSystem.out.println()).