Create area task
Developa  program to print the  area of a rectangle  by creating a  class named  Areahaving twoinstance variables as length and breadth of type double and twomethods. First method named as setDimension()takes length and breadth of rectangle as parameters and the second method named as getArea()returns the area of the rectangle. Length and breadth of rectangle are entered through keyboard.