Create Box task
Create a class called “Box” which contains three members: breadth, height and length of type double. Construct three objects of the “Box” as box1, box2 and box3. (5.0, 7.0, and 4.0) and (10.0, 12.0, 8.0) are the  specifications  for  box1  and  box2  respectively.  Build  a  Java  program  which add  box1  and  box2 resulting into box3, and display box3 coordinates on console.