Create CalculateArea TASK-4 OOP.java
Create a class to print the area of a square and a rectangle. The
class has two methods with the same name but different number of
parameters. The method for printing area of rectangle has two
parameters which are length and breadth respectively while the
other method for printing area of square has one parameter which is
side of square.