Create Student Task-5 OOP.java
Create a class 'Student' with three data members which are name,
age and address. The constructor of the class assigns default values
name as "unknown", age as '0' and address as "not available". It has
two members with the same name 'setInfo'. First method has two
parameters for name and age and assigns the same whereas the
second method takes has three parameters which are assigned to
name, age and address respectively. Print the name, age and
address of 10 students.
Hint - Use array of objects