Create 5.1.java
/*****
Define a class for rectangle objects defined by two points,
the top-left and bottom-right corners of the rectangle.
Include a constructor to copy a rectangle,
a method to return a rectangle object that encloses the current object and the rectangle passed as an argument,
and a method to display the defining points of a rectangle. 
Test the class by creating four rectangles and combining these cumulatively
to end up with a rectangle enclosing them all. Output the defining points of all the rectangles you create.
******/